abstracted out fetch calls that require the csrf token, replaced duplicate functions in actions & in the DeleteElementButton component
